Refactor load Products creating a custom hook
